Test: Avoid Caddy redirect for index page
Let the Jenkins test instance pretend to be a Jenkins instance which has
htmlpublisher-wrapper.html as index page. The Caddy default is to
consider index.html as default which does not correspond with the live
environment. Additionally, recent versions of Caddy started redirecting
index.html URLs to the URL without index.html, which causes the NGINX
reverse proxy to start redirecting the rewritten URL again, leading the
selenium browser to stop redirecting after a while.