Test: Downgrade selenium and python runner for now
Selenium 4 is having some weird slowness at the end of each test where
it hangs on ending the session, which makes each test take five minutes
longer. https://github.com/SeleniumHQ/docker-selenium/issues/1503 may
indicate a possible related issue.

Python 3.8 is slow to set up since it does not have a musl-supported
binary wheel for the cffi package and thus needs to be compiled with gcc
and headers. https://foss.heptapod.net/pypy/cffi/-/issues/509 indicates
possible progress toward making it available. 3.6 does not use
cryptography and cffi (yet?). Other solutions are not using Alpine for
this image, to build and tag the Docker as a downstream image containing
the packages, or to use the system Python (like the rest of the Jenkins
pipeline already does) which then connects to the docker compose network
(means selenium port must be exposed to the host).

For now, downgrade works best.