Correct CSS parsing problems in Firefox
The CSS declaration trick that avoids skipping several declarations when
using a pseudo-element/selector (by adding a selector for all namespaces
and all elements *|* in front of it) is being stripped out in production
builds causing difference in behavior between production and development
(which could not even be noticed in the test setup even if using Firefox
for the Selenium tests). Firefox is kind of buggy in this regard but the
stripping of the namespace/element selector appears to be an intentional
function of the laravel-mix/webpack/node-sass/dart-sass/postcss compiler
stack. By getting rid of the namespace trick and letting the selector be
handled by autoprefixer appears to resolve the issue, making other rules
such as full-sized navbar icons and dropdown header icon margins come to
effect in production like in development, and this is more future-proof.