Navbar: Remove burger in fullscreen mode
Fullscreen mode on mobile can be useful to make use of the full device
screen, so displaying the button as a floating element is practical.
However, the burger remained visible in fullscreen, an oversight since
all other elements were made invisible, thus the burger did nothing.
Hide the burger in fullscreen mode on mobile to align with desktop.