Apache: Add per-filetype CORS header
Somehow the "Header set" based on environment variable does not work,
even outside LocationMatch (order of rule processing for LocationMatch
rules?), instead use a FilesMatch to allow access to these remotely.
This is mostly for development where loading in the "visualization-site"
resources is helpful, but may also be used in production if
visualization and prediction are on different hostnames, and does not
introduce security issues.